+++ CHANGELOG.md
@@ -7,14 +7,19 @@ SneakyTactician  <SneakyTactician@outlook.com>
 #### API
 * World data is now structured differently, and hopefully more efficiently
 
+#### Server
+* World generation now no longer spawns the character on a resource
+
 #### Bugs
 * Discovered and fixed issue [#21](https://github.com/SneakyTactician/MagicalLife/issues/21)
 * Discovered and fixed issue [#22](https://github.com/SneakyTactician/MagicalLife/issues/22)
 * Discovered and fixed issue [#24](https://github.com/SneakyTactician/MagicalLife/issues/24)
-* Fixed issue [#11](https://github.com/SneakyTactician/MagicalLife/issues/11)
-* Fixed issue [#18](https://github.com/SneakyTactician/MagicalLife/issues/18)
 * Discovered and fixed issue [#25](https://github.com/SneakyTactician/MagicalLife/issues/25)
 
+* Fixed issue [#11](https://github.com/SneakyTactician/MagicalLife/issues/11)
+* Fixed issue [#18](https://github.com/SneakyTactician/MagicalLife/issues/18)
+
+* Discovered issue [#26](https://github.com/SneakyTactician/MagicalLife/issues/26)
 ---
 ## [Version 0.0.7]
 #### GUI